Move configure logging into share script
