+++ tools/build_release.py
@@ -1,30 +1,16 @@
 # !python
 
 import argparse
+from configure_logging import configure_logging
 import logging
 import pprint
 from subprocess import check_call
-import sys
 
 pp = pprint.PrettyPrinter(indent=4, width=100)
 
 default_python_cmd = ['python.exe']
 drivers_to_upload = ['nidcpower', 'nidigital', 'nidmm', 'niswitch', 'nimodinst', 'nifgen', 'niscope', 'nise', 'nitclk']
 drivers_to_update = ['nifake'] + drivers_to_upload
-
-
-def configure_logging(lvl=logging.WARNING, logfile=None):
-    root = logging.getLogger()
-    root.setLevel(lvl)
-
-    formatter = logging.Formatter("[%(asctime)s] [%(levelname)8s] --- %(message)s (%(filename)s:%(funcName)s:%(lineno)s)", "%Y-%m-%d %H:%M:%S")
-    if logfile is None:
-        hndlr = logging.StreamHandler(sys.stdout)
-    else:
-        print("Logging to file %s" % logfile)
-        hndlr = logging.FileHandler(logfile)
-    hndlr.setFormatter(formatter)
-    root.addHandler(hndlr)
 
 
 class CustomFormatter(argparse.ArgumentDefaultsHelpFormatter, argparse.RawDescriptionHelpFormatter):
